refactor(auth): reorganize auth dtos and implement use case interfaces
Move auth-related DTOs to dedicated auth directory and implement interfaces for use cases to improve type safety and maintainability
