+++ services/user-admin-service/src/application/usecases/admin/LoginAdminUseCase.ts
@@ -2,7 +2,7 @@ import bcrypt from "bcrypt";
 
 import { tokenService } from "../../../presentation/service/token.service";
 import { AuthResponseDto } from "../../../domain/dtos/user/AuthResponseDto";
-import { LoginRequestDto } from "../../../domain/dtos/user/LoginRequestDto";
+import { LoginRequestDto } from "../../../domain/dtos/user/auth/LoginRequestDto";
 import { IUserRepository } from "../../../domain/repositories/user/IUserRepository";
 import {
   ERROR_MESSAGES,
